+++ src/gui/tasks.rs
@@ -392,7 +392,10 @@ pub fn import_cemu_settings(core: &Manager, path: &Path) -> Result<Message> {
         let deploy_config = wiiu_config.deploy_config.get_or_insert_default();
         deploy_config.auto = true;
         deploy_config.output = gfx_folder.join("BreathOfTheWild_UKMM");
-        deploy_config.executable = gfx_folder.with_file_name("Cemu.exe").exists_then();
+        deploy_config.executable = gfx_folder
+            .with_file_name("Cemu.exe")
+            .exists_then()
+            .map(|p| p.display().to_string());
     } else {
         settings.wiiu_config = Some(PlatformSettings {
             language: uk_content::constants::Language::USen,
@@ -403,7 +406,10 @@ pub fn import_cemu_settings(core: &Manager, path: &Path) -> Result<Message> {
                 method: uk_manager::settings::DeployMethod::Symlink,
                 output: gfx_folder.join("BreathOfTheWild_UKMM"),
                 cemu_rules: true,
-                executable: gfx_folder.with_file_name("Cemu.exe").exists_then(),
+                executable: gfx_folder
+                    .with_file_name("Cemu.exe")
+                    .exists_then()
+                    .map(|p| p.display().to_string()),
             }),
         })
     };