+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added support for new Cemu config paths
 - Experimental support for arm64 macOS
+- Added support for complex emulator command lines
 
 **Changed**